+++ build_resolvers/test/resolver_test.dart
@@ -82,6 +82,24 @@ void runTests(ResolversFactory resolversFactory) {
     }, resolvers: createResolvers());
   });
 
+  test('does not stack overflow on long import chain', () {
+    return resolveSources(
+      {
+        'a|web/main.dart': '''
+              import 'lib0.dart';
+
+              main() {
+              } ''',
+        for (var i = 0; i != 750; ++i)
+          'a|web/lib$i.dart': i == 749 ? '' : 'import "lib${i + 1}.dart";',
+      },
+      (resolver) async {
+        await resolver.libraryFor(entryPoint);
+      },
+      resolvers: createResolvers(),
+    );
+  });
+
   test('should follow package imports', () {
     return resolveSources({
       'a|web/main.dart': '''
@@ -746,8 +764,9 @@ int? get x => 1;
             resolver.compilationUnitFor(AssetId.parse('a|errors.dart')),
             throwsA(isA<SyntaxErrorInAssetException>()),
           );
-        });
+        }, resolvers: createResolvers());
       });
+
       test('are only reported if severe', () {
         return resolveSources({
           'a|errors.dart': '''
@@ -763,7 +782,7 @@ int? get x => 1;
             resolver.compilationUnitFor(AssetId.parse('a|errors.dart')),
             completion(isNotNull),
           );
-        });
+        }, resolvers: createResolvers());
       });
 
       test('are reported for part files with errors', () {