+++ build_test/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.2.4-wip
 
-- Support checks on reader state after a build acion in `resolveSources`.
+- Support checks on reader state after a build action in `resolveSources`.
+- Start using `package:build/src/internal.dart`.
 
 ## 2.2.3